lmnet/requirements.txt
beautifulsoup4==4.6.0
click==6.7
easydict==1.6
matplotlib==2.2.2
numpy==1.14.0
pandas==0.19.2
Pillow==3.1.2
pytablewriter==0.35
PyYAML==3.12
